WIP: Specialized tokenizer
I went down the path of being able to tokenize a template using
specialized tokenizers for the specific platform.

Once I got into tokenizing lines that needed context (i.e. what platform
are they for), then it started getting complicated.

It was dealing with lines one at a time but it really should be dealing
with tokens (one character at a time) in order to make multi-line
elements and logic easier.

I think the next step is to create a more traditional Tokenizer so it
can be passed to platform-specific parsers. The tokenizer ensures that
the template structure is correct and it should not infer meaning about
the tokens. That is the job of the parser.